+++ src/IdentityPolicy.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace CharlieLangridge\Playa;
+
+enum IdentityPolicy: string
+{
+    case Session = 'session';
+    case Rolling = 'rolling';
+}